Placing and initing test correctly
